ADD issue

+++ src/main/java/Baekjoon1008.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Baekjoon1008 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int A = sc.nextInt();
-        int B = sc.nextInt();
-        System.out.println((double) A / B);
+        float A = sc.nextFloat();
+        float B = sc.nextFloat();
+        System.out.println(A / B);
         sc.close();
     }
 }
